requirements.txt
torch
torchaudio
transformers>=4.52
accelerate
huggingface_hub
sounddevice
numpy
soundfile
librosa
sentencepiece
protobuf
pyobjc-framework-Cocoa
pyobjc-framework-Quartz
pyyaml
